agent/requirements.txt
openai
python-dotenv
tavily-python
requests
beautifulsoup4
pytest
